// frontend/src/types/song.ts
export type SongType = {
  TrackId: string;
  ArtistName: string;
  TrackName: string;
  Popularity: number;
  Year: number;
  Genre: string;
  Danceability: number;
  Energy: number;
  Key: number;
  Loudness: number;
  Mode: number;
  Speechiness: number;
  Acousticness: number;
  Intrumentalness: number;
  Liveness: number;
  Valence: number;
  Tempo: number;
  DurationMs: number;
  TimeSignature: number;
};